Docker and validation

+++ backend/src/main/java/com/codrut/probalsac/review/service/ReviewService.java
@@ -5,11 +5,14 @@ import com.codrut.probalsac.review.controller.dto.ReviewDTO;
 import com.codrut.probalsac.review.controller.dto.UpdateReviewDTO;
 import com.codrut.probalsac.review.mapper.ReviewMapper;
 import com.codrut.probalsac.review.repository.ReviewRepository;
+import com.codrut.probalsac.user.domain.User;
 import com.codrut.probalsac.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,8 +24,7 @@ public class ReviewService {
     private final ReviewMapper mapper;
 
     public void save(ReviewCreationDTO dto) {
-        var user = userRepository.findById(dto.user_id)
-                .orElseThrow(() -> new RuntimeException("The user couldn't be found!"));
+        var user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         var entity = mapper.mapToEntity(dto, user);
 
         repository.save(entity);
@@ -43,7 +45,13 @@ public class ReviewService {
     }
 
     public void update(Long id, UpdateReviewDTO dto) {
+        var user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         var entity = repository.getById(id);
+
+        if(!Objects.equals(entity.getUser().getId(), user.getId())) {
+            throw new RuntimeException("The user doesn't have permission for this");
+        }
 
         if (dto.message != null)
             entity.setMessage(dto.message);
@@ -52,6 +60,13 @@ public class ReviewService {
     }
 
     public void delete(Long id) {
+        var user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        var entity = repository.getById(id);
+
+        if(!Objects.equals(entity.getUser().getId(), user.getId())) {
+            throw new RuntimeException("The user doesn't have permission for this");
+        }
+
         repository.deleteById(id);
     }
