+++ backend/src/main/java/com/codrut/probalsac/user/service/UserService.java
@@ -4,9 +4,12 @@ import com.codrut.probalsac.user.controller.dto.RegisterUserDTO;
 import com.codrut.probalsac.user.controller.dto.UpdateUserDTO;
 import com.codrut.probalsac.user.controller.dto.UserDTO;
 import com.codrut.probalsac.user.domain.Role;
+import com.codrut.probalsac.user.domain.User;
 import com.codrut.probalsac.user.mapper.UserMapper;
 import com.codrut.probalsac.user.repository.UserRepository;
+import com.codrut.probalsac.user.validation.UserValidation;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -14,6 +17,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,6 +27,7 @@ public class UserService implements UserDetailsService {
     private final UserRepository repository;
     private final UserMapper mapper;
     private final PasswordEncoder passwordEncoder;
+    private final UserValidation validation;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
@@ -33,7 +38,7 @@ public class UserService implements UserDetailsService {
     public void save(RegisterUserDTO dto) {
         var entity = mapper.mapToEntity(dto);
 
-        //TODO: validation
+        validation.validate(dto);
         entity.setPassword(passwordEncoder.encode(entity.getPassword()));
 
         repository.save(entity);
@@ -55,6 +60,11 @@ public class UserService implements UserDetailsService {
 
     public void update(Long id, UpdateUserDTO dto) {
         var userToUpdate = repository.getById(id);
+        var user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!Objects.equals(user.getId(), id)) {
+            throw new RuntimeException("The user doesn't have permission for this");
+        }
 
         if (dto.firstname != null && !dto.firstname.isEmpty())
             userToUpdate.setFirstname(dto.firstname);
@@ -75,6 +85,11 @@ public class UserService implements UserDetailsService {
     }
 
     public void delete(Long id) {
+        var user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!Objects.equals(id, user.getId())) {
+            throw new RuntimeException("The user doesn't have permission for this");
+        }
         repository.deleteById(id);
     }
